# docs/swagger.yaml
basePath: /
definitions:
  models.ErrorResponse:
    properties:
      error:
        type: string
      errors:
        items:
          $ref: '#/definitions/models.ValidationError'
        type: array
    type: object
  models.PaymentRequest:
    properties:
      amount:
        type: integer
      card_number:
        type: string
      currency:
        type: string
      cvv:
        type: string
      expiry_month:
        type: integer
      expiry_year:
        type: integer
    type: object
  models.PaymentResponse:
    properties:
      amount:
        type: integer
      card_number_last_four:
        type: string
      currency:
        type: string
      expiry_month:
        type: integer
      expiry_year:
        type: integer
      id:
        type: string
      status:
        type: string
    type: object
  models.ValidationError:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
host: localhost:8090
info:
  contact: {}
  description: Interview challenge for building a Payment Gateway - Go version
  title: Payment Gateway Challenge Go
paths:
  /api/payments:
    post:
      consumes:
      - application/json
      description: Processes a card payment through the payment gateway
      parameters:
      - description: Payment Request
        in: body
        name: payment
        required: true
        schema:
          $ref: '#/definitions/models.PaymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PaymentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Process a payment
      tags:
      - payments
  /api/payments/{id}:
    get:
      description: Retrieves details of a previously made payment by its ID
      parameters:
      - description: Payment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PaymentResponse'
        "404":
          description: Not Found
      summary: Retrieve payment details
      tags:
      - payments
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"